refactor(eatinorder): OrderTable persistance 계층 서비스 외부로 분리
