+++ src/main/java/kitchenpos/eatinorder/adapter/in/rest/OrderTableRestController.java
@@ -3,7 +3,6 @@ package kitchenpos.eatinorder.adapter.in.rest;
 import kitchenpos.eatinorder.application.service.OrderTableService;
 import kitchenpos.eatinorder.application.service.model.ChangeNumberOfGuestsRequest;
 import kitchenpos.eatinorder.application.service.model.CreateOrderTableRequest;
-import kitchenpos.eatinorder.domain.model.OrderTableEntity;
 import kitchenpos.eatinorder.domain.model.todo.OrderTable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
