--- src/main/java/kitchenpos/eatinorder/domain/model/OrderType.java
@@ -1,5 +1,0 @@
-package kitchenpos.eatinorder.domain.model;
-
-public enum OrderType {
-    DELIVERY, TAKEOUT, EAT_IN
-}
